refactor: add CreateUser useCase
